Advanced-Game-News-Settings Initial Commit
Added AdvancedGameNewsSettings
Added NotableMissesSettings
Added Logic to handle notable misses, but still needs to be cleaned up as I accidentaly did the logic in non league area and should only be relevant to league settings

In all places that use the the old GameNewsSetting, Ive replaced it with the new AdvancedGameNewsSetting.

The old discord command will still act the same but will use the new AdvancedGameNewsSetting class in the backend.

- need to add a new discord command that lets you tune these settings granularly